requirements.txt
sqlacodegen==2.1.0

pip~=20.1
wheel~=0.34.2
setuptools~=46.1.3
Cython~=0.29.17
SQLAlchemy~=1.3.16
psycopg2~=2.8.5
zipp~=3.1.0

requests==2.23.0
requests-cache==0.5.2

python-slugify==4.0.0
